Some fixes here and there...
Also implemented serial console in userspace

+++ src/kernel/syscall/actions/sys_map.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <events/event.h>
 
-u32 sys_map(u32 id, u32 *function)
+u32 sys_map(u32 id, u8 *function)
 {
 	event_map(id, function);
 	return -1;
