+++ src/kernel/syscall/syscall.h
@@ -26,6 +26,6 @@ u32 sys_free(u32 ptr);
 
 u32 sys_get(u32 id);
 
-u32 sys_map(u32 id, u32 *function);
+u32 sys_map(u32 id, u8 *function);
 
 #endif
